Remove Home Agent Playground refactoring
